Fix action key casing, split spec file, fix pauseOffset comments
Agent-Logs-Url: https://github.com/rmartz/hidden-role-game/sessions/a353476b-3aa9-4e6f-b078-7c0f1585c693
